patch: upgrade skill generator factory moved service factory genorators can ask for confirmation before overwriting a file extra warning if tests have dangling prompts removed some ts-ignores nameReadable required only ask required questions

+++ packages/spruce-cli/src/generators/AbstractGenerator.ts
@@ -2,27 +2,33 @@ import fs from 'fs'
 import pathUtil from 'path'
 import { diskUtil } from '@sprucelabs/spruce-skill-utils'
 import { Templates } from '@sprucelabs/spruce-templates'
-import { IGeneratedFile } from '../types/cli.types'
-
-export interface IGeneratorOptions {
-	templates: Templates
-}
+import { IGeneratedFile, IGraphicsInterface } from '../types/cli.types'
 
 export type GenerationResults = IGeneratedFile[]
 
+export interface IGeneratorOptions {
+	templates: Templates
+	term: IGraphicsInterface
+	askBeforeUpdating?: boolean
+}
+
 export default abstract class AbstractGenerator {
 	protected templates: Templates
+	private askBeforeUpdating = false
+	private term: IGraphicsInterface
 
-	public constructor(templates: Templates) {
-		this.templates = templates
+	public constructor(options: IGeneratorOptions) {
+		this.templates = options.templates
+		this.term = options.term
+		this.askBeforeUpdating = !!options.askBeforeUpdating
 	}
 
-	protected writeFileIfChangedMixinResults(
+	protected async writeFileIfChangedMixinResults(
 		destination: string,
 		contents: string,
 		description: string,
 		results?: GenerationResults
-	): GenerationResults {
+	): Promise<GenerationResults> {
 		const myResults: GenerationResults = results ?? []
 
 		const name = pathUtil.basename(destination)
@@ -32,8 +38,15 @@ export default abstract class AbstractGenerator {
 			diskUtil.writeFile(destination, contents)
 			action = 'generated'
 		} else if (diskUtil.isFileDifferent(destination, contents)) {
-			diskUtil.writeFile(destination, contents)
-			action = 'updated'
+			let write = true
+			if (this.askBeforeUpdating) {
+				write = await this.term.confirm(`Overwrite ${destination}?`)
+			}
+
+			if (write) {
+				diskUtil.writeFile(destination, contents)
+				action = 'updated'
+			}
 		}
 
 		myResults.push({ name, description, path: destination, action })
